+++ src/Settings/PreferencesPageTemplate.cpp
@@ -77,6 +77,13 @@ PreferencesPageTemplate::PreferencesPageTemplate(QStringList opts, bool alignTop
             addRow(wrapper, si.tip, si.help, si.desc);
             widgets.push_back(wrapper);
         }
+        else if (si.type == "QList<QVariant>")
+        {
+            Wrapper<QList<QVariant>> *wrapper = createStringListsWrapper(si.ui);
+            wrapper->init(this, si.param);
+            addRow(wrapper, si.tip, si.help, si.desc);
+            widgets.push_back(wrapper);
+        }
     }
 }