+++ doc/CHANGELOG.md
@@ -8,6 +8,9 @@ Now error linting is available with Language Server. Linting can be helpful when
 
 - Add Ctrl+Shift+D for duplicating the current line or the selection. (#237)
 - Add tooltips for many widgets. (#211)
+- Now you can add pairs of test cases from files. (#204 and #246)
+- Now you can choose where to save the test case files in Preferences->File Path->Testcases. (#176)
+- Now when saving the source file, if the parent directory of the file does not exist, it will be automatically created.
 
 ### Changed